Reverted HashSet change. Removed package name.
An ordered collection still fulfills the requirement of being unordered.
In addition, duplicates are not forbidden in the peers list, which using
a Set would block.

No package name is defined in the spec, so I've moved them out of the
package src, regardless of how bad a practice that is.